Reorganise Theme page view
