+++ content/webapp/pages/concepts/[conceptId].tsx
@@ -1,37 +1,11 @@
 import { GetServerSideProps, NextPage } from 'next';
-import { LinkProps } from 'next/link';
-import { usePathname } from 'next/navigation';
-import { FunctionComponent, JSX, useState } from 'react';
-import styled from 'styled-components';
 
-import { pageDescriptionConcepts } from '@weco/common/data/microcopy';
-import { ImagesLinkSource } from '@weco/common/data/segment-values';
 import { getServerData } from '@weco/common/server-data';
-import { useToggles } from '@weco/common/server-data/Context';
 import { appError, AppErrorProps } from '@weco/common/services/app';
 import { Pageview } from '@weco/common/services/conversion/track';
-import { font } from '@weco/common/utils/classnames';
-import { capitalize, formatNumber } from '@weco/common/utils/grammar';
 import { serialiseProps } from '@weco/common/utils/json';
-import { getQueryResults, ReturnedResults } from '@weco/common/utils/search';
+import { getQueryResults } from '@weco/common/utils/search';
 import { ApiToolbarLink } from '@weco/common/views/components/ApiToolbar';
-import { Container } from '@weco/common/views/components/styled/Container';
-import Space from '@weco/common/views/components/styled/Space';
-import { themeValues } from '@weco/common/views/themes/config';
-import theme from '@weco/common/views/themes/default';
-import CatalogueImageGallery from '@weco/content/components/CatalogueImageGallery';
-import CataloguePageLayout from '@weco/content/components/CataloguePageLayout';
-import MoreLink from '@weco/content/components/MoreLink';
-import { toLink as toImagesLink } from '@weco/content/components/SearchPagesLink/Images';
-import { toLink as toWorksLink } from '@weco/content/components/SearchPagesLink/Works';
-import Tabs from '@weco/content/components/Tabs';
-import ThemeCollaborators from '@weco/content/components/ThemeCollaborators';
-import ThemeHeader from '@weco/content/components/ThemeHeader';
-import ThemeImages from '@weco/content/components/ThemeImages';
-import ThemeRelatedConceptsGroup from '@weco/content/components/ThemeRelatedConceptsGroup';
-import ThemeWorks from '@weco/content/components/ThemeWorks';
-import WorksSearchResults from '@weco/content/components/WorksSearchResults';
-import useHotjar from '@weco/content/hooks/useHotjar';
 import { emptyResultList } from '@weco/content/services/wellcome';
 import { looksLikeCanonicalId } from '@weco/content/services/wellcome/catalogue';
 import { getConcept } from '@weco/content/services/wellcome/catalogue/concepts';
@@ -41,7 +15,6 @@ import {
   Concept as ConceptType,
   Image as ImageType,
   toWorkBasic,
-  WorkBasic,
   Work as WorkType,
 } from '@weco/content/services/wellcome/catalogue/types';
 import { getWorks } from '@weco/content/services/wellcome/catalogue/works';
@@ -52,443 +25,9 @@ import {
   queryParams,
 } from '@weco/content/utils/concepts';
 import { cacheTTL, setCacheControl } from '@weco/content/utils/setCacheControl';
-
-const emptyImageResults: CatalogueResultsList<ImageType> = emptyResultList();
-
-const emptyWorkResults: CatalogueResultsList<WorkType> = emptyResultList();
-
-const tabOrder = ['by', 'in', 'about'] as const;
-
-const linkSources = new Map([
-  ['worksAbout', 'concept/works_about'],
-  ['worksBy', 'concept/works_by'],
-  ['worksIn', 'concept/works_in'],
-  ['imagesAbout', 'concept/images_about'],
-  ['imagesBy', 'concept/images_by'],
-  ['imagesIn', 'concept/images_in'],
-]);
-
-const HotJarPlaceholder = styled.div`
-  margin: -2rem auto 2rem auto;
-  width: 100%;
-  max-width: ${themeValues.sizes.xlarge}px;
-
-  display: grid;
-  justify-items: start;
-  padding: 0 ${themeValues.containerPadding.small}px;
-
-  div:has(form) {
-    min-width: 250px;
-  }
-
-  grid-template-columns: 1fr auto;
-
-  ${themeValues.media('medium')(`
-    padding: 0 ${themeValues.containerPadding.medium}px;
-    div:has(form) {
-      min-width: 350px;
-    }
-  `)}
-
-  ${themeValues.media('large')(`
-    padding: 0 ${themeValues.containerPadding.large}px;
-    div:has(form) {
-      min-width: 450px;
-    }
-  `)}
-`;
-
-const ConceptHero = styled(Space).attrs({
-  $v: { size: 'l', properties: ['padding-top', 'padding-bottom'] },
-})`
-  background-color: ${props => props.theme.color('lightYellow')};
-`;
-
-const HeroTitle = styled.h1.attrs({ className: font('intb', 1) })`
-  margin-bottom: 1rem;
-`;
-
-const TypeLabel = styled.span.attrs({ className: font('intb', 6) })`
-  background-color: ${props => props.theme.color('warmNeutral.300')};
-  padding: 5px;
-`;
-
-const ConceptImages = styled(Space).attrs({
-  $v: { size: 'xl', properties: ['padding-top', 'padding-bottom'] },
-})`
-  background-color: ${props => props.theme.color('black')};
-
-  .sectionTitle {
-    color: ${props => props.theme.color('white')};
-  }
-`;
-
-type ConceptWorksHeaderProps = {
-  $hasWorksTabs: boolean;
-};
-const ConceptWorksHeader = styled(Space).attrs({
-  $v: { size: 'xl', properties: ['padding-top'] },
-})<ConceptWorksHeaderProps>`
-  background-color: ${({ $hasWorksTabs, theme }) =>
-    theme.color($hasWorksTabs ? 'warmNeutral.300' : 'white')};
-`;
-
-// tabDefinitions is an ordered list of the image or works tabs in a page.
-// (hence not just having an object and doing a [selectedTab] lookup)
-// Return the currently selected one.
-function currentTabPanel<T>(
-  selectedTab: string,
-  tabDefinitions: PageSectionDefinition<T>[]
-) {
-  // When only one tab exists, there is no selection going on, just return it.
-  if (tabDefinitions.length === 1) return tabDefinitions[0].panel;
-
-  for (const definition in tabDefinitions) {
-    if (tabDefinitions[definition].id === selectedTab) {
-      return tabDefinitions[definition].panel;
-    }
-  }
-  throw new Error(
-    `Unexpected selected tab ${selectedTab} not found in ${tabDefinitions}`
-  );
-}
-
-type SeeMoreButtonType = {
-  text: string;
-  link: LinkProps;
-  totalResults: number;
-};
-
-const SeeMoreButton = ({ text, link, totalResults }: SeeMoreButtonType) => (
-  <MoreLink
-    name={`${text} (${formatNumber(totalResults, {
-      isCompact: true,
-    })})`}
-    url={link}
-    colors={theme.buttonColors.yellowYellowBlack}
-  />
-);
-
-type TabLabelProps = {
-  text: string;
-  totalResults: number;
-};
-
-const TabLabel = ({ text, totalResults }: TabLabelProps) => (
-  <>
-    {text} <span className="is-hidden-s">({formatNumber(totalResults)})</span>
-  </>
-);
-
-type ImagesTabPanelProps = {
-  id: string;
-  link: LinkProps;
-  results: ReturnedResults<ImageType>;
-  totalResults: number | undefined;
-};
-const ImagesTabPanel: FunctionComponent<ImagesTabPanelProps> = ({
-  id,
-  link,
-  results,
-  totalResults,
-}) => {
-  return (
-    <div role="tabpanel" id={`tabpanel-${id}`} aria-labelledby={`tab-${id}`}>
-      <CatalogueImageGallery images={results.pageResults} variant="justified" />
-      <Space $v={{ size: 'm', properties: ['margin-top'] }}>
-        {!!totalResults && (
-          <SeeMoreButton
-            text="All images"
-            totalResults={totalResults}
-            link={link}
-          />
-        )}
-      </Space>
-    </div>
-  );
-};
-
-type WorksTabPanelProps = {
-  id: string;
-  link: LinkProps;
-  results: ReturnedResults<WorkBasic>;
-  totalResults: number | undefined;
-};
-const WorksTabPanel: FunctionComponent<WorksTabPanelProps> = ({
-  id,
-  link,
-  results,
-  totalResults,
-}) => {
-  return (
-    <Container>
-      <div role="tabpanel" id={`tabpanel-${id}`} aria-labelledby={`tab-${id}`}>
-        <WorksSearchResults works={results.pageResults} />
-        <Space $v={{ size: 'l', properties: ['padding-top'] }}>
-          {!!totalResults && (
-            <SeeMoreButton
-              text="All works"
-              totalResults={totalResults}
-              link={link}
-            />
-          )}
-        </Space>
-      </div>
-    </Container>
-  );
-};
-
-// Represents the data for a single tab/tab panel combination.
-type PageSectionDefinition<T> = {
-  id: string;
-  tab: {
-    id: string;
-    text: JSX.Element;
-  };
-  panel: {
-    id: string;
-    link: LinkProps;
-    results: ReturnedResults<T>;
-    totalResults: number | undefined;
-  };
-};
-type PageSectionDefinitionProps<T> = {
-  tabId: string;
-  resultsGroup: ReturnedResults<T> | undefined;
-  tabLabelText: string;
-  totalResults: number | undefined;
-  link: LinkProps;
-};
-
-function toPageSectionDefinition<T>({
-  tabId,
-  resultsGroup,
-  tabLabelText,
-  totalResults,
-  link,
-}: PageSectionDefinitionProps<T>): PageSectionDefinition<T> | undefined {
-  return resultsGroup?.totalResults
-    ? {
-        id: tabId,
-        tab: {
-          id: tabId,
-          text: TabLabel({
-            text: tabLabelText,
-            totalResults: resultsGroup.totalResults,
-          }),
-        },
-        panel: { id: tabId, link, results: resultsGroup, totalResults },
-      }
-    : undefined;
-}
-
-export type SectionData = {
-  label: string;
-  works: ReturnedResults<WorkBasic> | undefined;
-  images: ReturnedResults<ImageType> | undefined;
-  totalResults: { works: number | undefined; images: number | undefined };
-};
-
-export type ThemePageSectionsData = {
-  about: SectionData;
-  by: SectionData;
-  in: SectionData;
-};
-
-type Props = {
-  conceptResponse: ConceptType;
-  sectionsData: ThemePageSectionsData;
-  apiToolbarLinks: ApiToolbarLink[];
-  pageview: Pageview;
-};
-
-export const ConceptPage: NextPage<Props> = ({
-  conceptResponse,
-  sectionsData,
-  apiToolbarLinks,
-}) => {
-  useHotjar(true);
-  const { newThemePages, themePagesAllFields } = useToggles();
-
-  const pathname = usePathname();
-  const worksTabs = tabOrder
-    .map(relationship => {
-      const tabId = `works${capitalize(relationship)}`;
-
-      const data = sectionsData[relationship] as SectionData;
-
-      return toPageSectionDefinition<WorkBasic>({
-        tabId,
-        resultsGroup: data.works,
-        tabLabelText: data.label,
-        totalResults: data.totalResults.works,
-        link: toWorksLink(
-          allRecordsLinkParams(tabId, conceptResponse),
-          linkSources[tabId]
-        ),
-      });
-    })
-    .filter(e => !!e) as PageSectionDefinition<WorkBasic>[];
-
-  const hasWorks = worksTabs.length > 0;
-  const hasWorksTabs = worksTabs.length > 1;
-
-  const imagesTabs: PageSectionDefinition<ImageType>[] = tabOrder
-    .map(relationship => {
-      const tabId = `images${relationship
-        .charAt(0)
-        .toUpperCase()}${relationship.slice(1)}`;
-
-      return toPageSectionDefinition({
-        tabId,
-        resultsGroup: sectionsData[relationship].images,
-        tabLabelText: sectionsData[relationship].label,
-        totalResults: sectionsData[relationship].totalResults.images,
-        link: toImagesLink(
-          allRecordsLinkParams(tabId, conceptResponse),
-          `${linkSources[tabId]}_${pathname}` as ImagesLinkSource
-        ),
-      });
-    })
-    .filter(e => !!e) as PageSectionDefinition<ImageType>[];
-
-  const hasImages = imagesTabs.length > 0;
-  const hasImagesTabs = imagesTabs.length > 1;
-
-  // Set the default tab in each group to the first populated tab
-  // the two tabs lists are ordered consistently as defined by tabOrder,
-  // which is ordered so that the more specific tabs come first.
-  // Maximally one "more specific" tab is expected to be populated for any given concept.
-  const [selectedWorksTab, setSelectedWorksTab] = useState(
-    worksTabs[0]?.id || ''
-  );
-  const [selectedImagesTab, setSelectedImagesTab] = useState(
-    imagesTabs[0]?.id || ''
-  );
-
-  const { frequentCollaborators, relatedTopics } =
-    conceptResponse.relatedConcepts || {};
-
-  return (
-    <CataloguePageLayout
-      title={conceptResponse.label}
-      description={pageDescriptionConcepts(conceptResponse.label)}
-      url={{ pathname: `/concepts/${conceptResponse.id}`, query: {} }}
-      openGraphType="website"
-      siteSection="collections"
-      jsonLd={{ '@type': 'WebPage' }}
-      hideNewsletterPromo={true}
-      apiToolbarLinks={apiToolbarLinks}
-    >
-      {newThemePages && <ThemeHeader concept={conceptResponse} />}
-      {!newThemePages && (
-        <ConceptHero>
-          <Container>
-            <TypeLabel>{conceptTypeDisplayName(conceptResponse)}</TypeLabel>
-            <Space
-              $v={{ size: 's', properties: ['margin-top', 'margin-bottom'] }}
-            >
-              <HeroTitle>{conceptResponse.label}</HeroTitle>
-            </Space>
-          </Container>
-        </ConceptHero>
-      )}
-
-      {/* Images */}
-      {newThemePages && (
-        <ThemeImages sectionsData={sectionsData} concept={conceptResponse} />
-      )}
-      {!newThemePages && hasImages && (
-        <ConceptImages as="section" data-testid="images-section">
-          <Container>
-            <h2 className={`${font('wb', 3)} sectionTitle`}>Images</h2>
-            {hasImagesTabs && (
-              <Tabs
-                label="Images tabs"
-                tabBehaviour="switch"
-                items={imagesTabs.map(t => t.tab)}
-                selectedTab={selectedImagesTab}
-                setSelectedTab={setSelectedImagesTab}
-                isWhite
-                trackWithSegment
-              />
-            )}
-            <Space $v={{ size: 'l', properties: ['margin-top'] }}>
-              <ImagesTabPanel
-                {...currentTabPanel(selectedImagesTab, imagesTabs)}
-              />
-            </Space>
-          </Container>
-        </ConceptImages>
-      )}
-
-      {/* Works */}
-      {newThemePages && (
-        <ThemeWorks concept={conceptResponse} sectionsData={sectionsData} />
-      )}
-      {!newThemePages && hasWorks && (
-        <>
-          <ConceptWorksHeader $hasWorksTabs={hasWorksTabs}>
-            <Container>
-              <h2 className={font('wb', 3)}>Catalogue</h2>
-              {hasWorksTabs && (
-                <Tabs
-                  label="Works tabs"
-                  tabBehaviour="switch"
-                  selectedTab={selectedWorksTab}
-                  items={worksTabs.map(t => t.tab)}
-                  setSelectedTab={setSelectedWorksTab}
-                  trackWithSegment
-                  hideBorder
-                />
-              )}
-            </Container>
-          </ConceptWorksHeader>
-          <Space
-            as="section"
-            $v={{
-              size: 'xl',
-              properties: ['margin-top', 'margin-bottom'],
-            }}
-            data-testid="works-section"
-          >
-            <WorksTabPanel {...currentTabPanel(selectedWorksTab, worksTabs)} />
-          </Space>
-        </>
-      )}
-      {newThemePages &&
-        (conceptResponse.type === 'Person' || themePagesAllFields) && (
-          <Container>
-            <Space
-              $v={{ size: 'xl', properties: ['margin-top', 'margin-bottom'] }}
-            >
-              <ThemeCollaborators concepts={frequentCollaborators} />
-            </Space>
-            <Space
-              $v={{ size: 'xl', properties: ['margin-top', 'margin-bottom'] }}
-            >
-              <ThemeRelatedConceptsGroup
-                label="Related topics"
-                labelType="heading"
-                relatedConcepts={relatedTopics}
-                buttonColors={
-                  themeValues.buttonColors.pumiceTransparentCharcoal
-                }
-              />
-            </Space>
-          </Container>
-        )}
-      {
-        // This is a placeholder for the Hotjar embedded survey to be injected
-        // when the concept is a Person. It should be removed when the survey
-        // is no longer used.
-      }
-      {conceptResponse.type === 'Person' && (
-        <HotJarPlaceholder id="hotjar-embed-placeholder-concept-person" />
-      )}
-    </CataloguePageLayout>
-  );
-};
+import ConceptPage, {
+  ConceptPageProps,
+} from '@weco/content/views/concepts/concept';
 
 function createApiToolbarLinks(concept: ConceptType): ApiToolbarLink[] {
   const apiUrl = `https://api.wellcomecollection.org/catalogue/v2/concepts/${concept.id}`;
@@ -514,6 +53,24 @@ function createApiToolbarLinks(concept: ConceptType): ApiToolbarLink[] {
 
   return [apiLink, ...identifiers];
 }
+
+type Props = ConceptPageProps & {
+  pageview: Pageview;
+};
+
+export const Page: NextPage<Props> = ({
+  conceptResponse,
+  sectionsData,
+  apiToolbarLinks,
+}) => {
+  return (
+    <ConceptPage
+      conceptResponse={conceptResponse}
+      sectionsData={sectionsData}
+      apiToolbarLinks={apiToolbarLinks}
+    />
+  );
+};
 
 export const getServerSideProps: GetServerSideProps<
   Props | AppErrorProps
@@ -581,6 +138,9 @@ export const getServerSideProps: GetServerSideProps<
   const worksAboutPromiseByLabel = getConceptDocs.works.byLabel('worksAbout');
   const imagesAboutPromiseByLabel =
     getConceptDocs.images.byLabel('imagesAbout');
+
+  const emptyImageResults: CatalogueResultsList<ImageType> = emptyResultList();
+  const emptyWorkResults: CatalogueResultsList<WorkType> = emptyResultList();
 
   // Only Genres can have works or images "in" them
   // so don't bother executing this query for other types
@@ -782,4 +342,4 @@ export const getServerSideProps: GetServerSideProps<
   };
 };
 
-export default ConceptPage;
+export default Page;
