+++ content/webapp/views/concepts/concept/components/Collaborators.tsx
@@ -3,7 +3,6 @@ import styled from 'styled-components';
 
 import { organisation, user } from '@weco/common/icons';
 import { font } from '@weco/common/utils/classnames';
-import { Grid, GridCell } from '@weco/common/views/components/styled/Grid';
 import CollaboratorCard from '@weco/content/components/CollaboratorCard';
 import {
   ConceptType,
@@ -11,10 +10,6 @@ import {
 } from '@weco/content/services/wellcome/catalogue/types';
 
 const COLLABORATOR_COUNT_LIMIT = 3;
-
-const Heading = styled.h2.attrs({
-  className: font('intsb', 2),
-})``;
 
 const CollaboratorsWrapper = styled.div`
   display: flex;
@@ -25,23 +20,21 @@ const CollaboratorsWrapper = styled.div`
   ${props => props.theme.media('medium')(`flex-direction: row;`)}
 `;
 
-type Props = {
-  concepts?: RelatedConcept[];
-};
-
 const iconFromConceptType = (type: ConceptType) => {
   if (type === 'Organisation') return organisation;
   return user;
 };
 
-const ThemeCollaborators: FunctionComponent<Props> = ({ concepts }) => {
+const Collaborators: FunctionComponent<{
+  concepts?: RelatedConcept[];
+}> = ({ concepts }) => {
   if (!concepts || concepts.length === 0) {
     return null;
   }
 
   return (
     <>
-      <Heading>Frequent collaborators</Heading>
+      <h2 className={font('intsb', 2)}>Frequent collaborators</h2>
       <CollaboratorsWrapper>
         {concepts.slice(0, COLLABORATOR_COUNT_LIMIT).map(concept => (
           <CollaboratorCard
@@ -56,4 +49,4 @@ const ThemeCollaborators: FunctionComponent<Props> = ({ concepts }) => {
   );
 };
 
-export default ThemeCollaborators;
+export default Collaborators;
