+++ content/webapp/views/concepts/concept/components/Header.tsx
@@ -6,10 +6,9 @@ import { font } from '@weco/common/utils/classnames';
 import { capitalize } from '@weco/common/utils/grammar';
 import { Container } from '@weco/common/views/components/styled/Container';
 import Space from '@weco/common/views/components/styled/Space';
-import ThemeAlternativeLabels from '@weco/content/components/ThemeHeader/ThemeAlternativeLabels';
-import ThemeRelatedConceptsGroup from '@weco/content/components/ThemeRelatedConceptsGroup';
-import ThemeSourcedDescription from '@weco/content/components/ThemeSourcedDescription';
+import SourcedDescription from '@weco/content/components/SourcedDescription';
 import { Concept } from '@weco/content/services/wellcome/catalogue/types';
+import { RelatedConceptsGroup } from '@weco/content/views/concepts/concept';
 
 const ConceptHero = styled(Space).attrs({
   $v: { size: 'xl', properties: ['padding-top', 'padding-bottom'] },
@@ -28,11 +27,46 @@ const ThemeDescription = styled(Space).attrs({
   $v: { size: 'l', properties: ['margin-bottom'] },
 })``;
 
-type Props = {
-  concept: Concept;
+const AlternativeLabels = styled(Space).attrs({
+  className: font('intr', 6),
+  $v: { size: 'm', properties: ['margin-bottom'] },
+})`
+  display: flex;
+  flex-wrap: wrap;
+  gap: ${props => props.theme.spacingUnits['3']}px;
+  color: ${props => props.theme.color('neutral.700')};
+`;
+
+const AlternativeLabel = styled.span.attrs({
+  className: font('intr', 6),
+})`
+  border-right: 1px solid ${props => props.theme.color('neutral.700')};
+  padding-right: ${props => props.theme.spacingUnits['3']}px;
+
+  &:last-of-type {
+    border-right: 0;
+  }
+`;
+
+const ThemeAlternativeLabels: FunctionComponent<{
+  alternativeLabels?: string[];
+}> = ({ alternativeLabels }) => {
+  if (!alternativeLabels || alternativeLabels.length === 0) {
+    return null;
+  }
+
+  return (
+    <AlternativeLabels>
+      {alternativeLabels.map(label => (
+        <AlternativeLabel key={label}>{capitalize(label)}</AlternativeLabel>
+      ))}
+    </AlternativeLabels>
+  );
 };
 
-const ThemeHeader: FunctionComponent<Props> = ({ concept }) => {
+const ThemeHeader: FunctionComponent<{
+  concept: Concept;
+}> = ({ concept }) => {
   const { themePagesAllFields } = useToggles();
 
   const { narrowerThan, fieldsOfWork, people, relatedTo, broaderThan } =
@@ -49,7 +83,7 @@ const ThemeHeader: FunctionComponent<Props> = ({ concept }) => {
         )}
         {concept.description && (
           <ThemeDescription>
-            <ThemeSourcedDescription
+            <SourcedDescription
               description={capitalize(concept.description.text)}
               source={concept.description.sourceLabel}
               href={concept.description.sourceUrl}
@@ -58,7 +92,7 @@ const ThemeHeader: FunctionComponent<Props> = ({ concept }) => {
         )}
         <>
           {(concept.type === 'Person' || themePagesAllFields) && (
-            <ThemeRelatedConceptsGroup
+            <RelatedConceptsGroup
               label="Field of work"
               labelType="inline"
               relatedConcepts={fieldsOfWork}
@@ -66,22 +100,22 @@ const ThemeHeader: FunctionComponent<Props> = ({ concept }) => {
           )}
           {themePagesAllFields && (
             <>
-              <ThemeRelatedConceptsGroup
+              <RelatedConceptsGroup
                 label="Part of"
                 labelType="inline"
                 relatedConcepts={narrowerThan}
               />
-              <ThemeRelatedConceptsGroup
+              <RelatedConceptsGroup
                 label="Notable people in this field"
                 labelType="heading"
                 relatedConcepts={people}
               />
-              <ThemeRelatedConceptsGroup
+              <RelatedConceptsGroup
                 label="Related to"
                 labelType="heading"
                 relatedConcepts={relatedTo}
               />
-              <ThemeRelatedConceptsGroup
+              <RelatedConceptsGroup
                 label="Broader than"
                 labelType="heading"
                 relatedConcepts={broaderThan}
