+++ content/webapp/views/concepts/concept/index.tsx
@@ -1,0 +1,17 @@
+import Collaborators from './components/Collaborators';
+import Header from './components/Header';
+import ImagesResults from './components/ImagesResults';
+import RelatedConceptsGroup from './components/RelatedConceptsGroup';
+import WorksResults from './components/WorksResults';
+import ConceptPage from './view';
+
+export default ConceptPage;
+export {
+  Collaborators,
+  Header,
+  ImagesResults,
+  WorksResults,
+  RelatedConceptsGroup,
+};
+
+export type { Props as ConceptPageProps } from './view';
